refactor and edd error message. Getting auth error
